fix(workflow): reconcile EvaluateOnly marking with the operation ledger rebase
main landed #3447's caller-owned DeferOperationMark mechanism for
on_children_completed after this branch's merge base, giving that call site
an OperationID alongside EvaluateOnly for the first time. Register it in the
AC-EO-15 pin test (its own commit-then-mark bracket already satisfies the
same contract, via the newer flag) and split engine_test.go's new subtests
into their own file to stay under the 800-effective-line lint limit after
#3447's own +81 lines pushed it over.
